requirements.txt
flask==3.1.0
requests>=2.31.0
pywebview>=5.0
